+++ src/builtin.h
@@ -1,0 +1,11 @@
+// builtin.h
+#ifndef BUILTIN_H
+#define BUILTIN_H
+
+int PSH_EXIT(char **);
+int PSH_CD(char **);
+int PSH_ECHO(char **);
+int PSH_PWD(char **);
+int PSH_HISTORY(char **);
+
+#endif
